remove unused parameters from OSAbstraction.CTextContext; fixed NRE if trying to render a node without content; added TextContextCache; restoring controls: application running but not rendered correctly
