+++ src/ImGui/OSImplentation/Shared/Text/GlyphCache.cs
